+++ app/Http/Controllers/ProductController.php
@@ -5,9 +5,23 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use Validator;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 class ProductController extends Controller
 {
+
+
+
+
+    /**
+     * Create a new AuthController instance.
+     * 
+     * @return void
+     */
+    public function __construct() {
+        $this->middleware('api');
+    }
+ 
     /**
      * Display a listing of the resource.
      *
@@ -15,6 +29,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        
         $users = Product::all();
   return response()->json(['message' => 'Success','data'=>$users]);
     }
@@ -35,7 +50,16 @@ class ProductController extends Controller
      */
     public function create(Request $request)
     {
-        
+        $user = Auth::guard('api')->user();
+
+if($user->is_admin!=1)
+{
+      return response()->json([
+              'message' => 'Unauthenticated',
+            
+            
+        ]);
+}
         if(Product::create($request->all()))
         {
             return response()->json(['message' => 'Success']);
@@ -79,6 +103,16 @@ class ProductController extends Controller
      */
     public function update(Request $request)
     {
+        $user = Auth::guard('api')->user();
+
+if($user->is_admin!=1)
+{
+      return response()->json([
+              'message' => 'Unauthenticated',
+            
+            
+        ]);
+}
         $product = Product::find($request->id);
          $validator = Validator::make($request->all(), [
             'name' => 'required|string|between:2,100',
@@ -118,6 +152,16 @@ else
      */
     public function destroy(Request $request)
     {
+        $user = Auth::guard('api')->user();
+
+if($user->is_admin!=1)
+{
+      return response()->json([
+              'message' => 'Unauthenticated',
+            
+            
+        ]);
+}
         $product = Product::find($request->id);
        
         if($product->delete())
